fixed linting errors; added ENABLED_ENDPOINTS to config; renamed api_util to mime_type_util
