+++ src/settings-example.py
@@ -12,6 +12,8 @@ class Config(object):
 
     STORAGE_BASE_URL = 'http://prd-app-bng-01.beeldengeluid.nl:8101/'
 
+    ENABLED_ENDPOINTS = ['resource', 'concept', 'dataset'] # allow all by default
+
     PROFILES = [
         {
             'title': 'NISV Catalogue schema',
